refactor(edge): reduce quic listener module to runtime facade
