+++ crates/edge/src/quic_listener/control_api/http.rs
@@ -1,4 +1,8 @@
+use std::sync::atomic::AtomicUsize;
+
 use ::http::{Method, header};
+use bytes::Bytes;
+use http_body_util::Full;
 
 use super::{
     state::{ConnectionSlotGuard, ControlApiListenerBinding, ControlApiPaths, ControlApiState},
