+++ crates/edge/src/quic_listener/control_api/state.rs
@@ -1,3 +1,5 @@
+use std::sync::atomic::AtomicUsize;
+
 use super::*;
 
 #[derive(Clone)]
