+++ crates/edge/src/quic_listener/forwarding/auth.rs
@@ -1,5 +1,7 @@
 use std::convert::Infallible;
+use std::sync::OnceLock;
 
+use http_body_util::Full;
 use hyper_rustls::HttpsConnectorBuilder;
 use hyper_util::client::legacy::{Client, connect::HttpConnector};
 use serde_json::Value;
