+++ crates/edge/src/quic_listener/forwarding/prepare.rs
@@ -1,5 +1,6 @@
 use std::convert::Infallible;
 
+use http_body_util::Full;
 use spooky_config::runtime::RuntimeExternalAuth;
 use tracing::Span;
